Add handshake codec
The goal of the this codec is to serialize `params.UpgradeConfig` in a
deterministic way, to be hashed later. This hash is going to be share by nodes
at handshake, so they can determine if they have the same upgrade config.

This PR leverages the newly introduced support to serilize Maps
https://github.com/ava-labs/avalanchego/pull/1790, which are deterministic (the
values are sorted by keys before serializing)